requirements.txt
# langChain 相关包
langchain==0.3.27
langchain_ollama==0.3.6
langchain_community==0.3.27
requests==2.32.4
PyPDF2==3.0.1
python-docx==1.2.0
unstructured==0.18.11
transformers==4.54.0
sentence-transformers==5.0.0
langchain-openai==0.3.28
dashscope==1.24.0
gradio==5.42.0
torch==2.7.1
uv==0.8.4
mcp==1.12.2
playwright==1.54.0
langchain-experimental==0.3.4
pandas-2.3.1
langchain-tavily-0.2.11
playwright==1.54.0
lxml==6.0.0
beautifulsoup4==4.13.4
reportlab==4.4.3
langchain-mcp-adapters==0.1.9
nest_asyncio==1.6.0
pyaudio==0.2.14
pydub==0.25.1
playsound==1.3.0
webrtcvad==2.0.10
langgraph==0.6.4
langsmith==0.4.13
langchain-openai==0.3.30
scipy==1.16.1
streamlit==1.48.1
faiss-cpu==1.12.0
numexpr==2.11.0


# 模型下载
modelscope==1.29.2